fixed most of bugs, added a spell system (not entirely done, missing hud and cd on spell change and many more things), continued to reformat code

+++ Damage.cs
@@ -25,7 +25,7 @@ namespace Futuridium
 
         public Vector2 InverseDirection { get; private set; }
 
-        public Func<Character, Character, int> DamageFunc { private get; set; }
+        public Func<Character, Character, float> DamageFunc { private get; set; }
 
         public Character Enemy { get; set; }
 
@@ -33,7 +33,7 @@ namespace Futuridium
 
         public Character Character { get; set; }
 
-        public int Caculate(Character character, Character enemy)
+        public float Caculate(Character character, Character enemy)
         {
             return DamageFunc(character, enemy);
         }
